Update the run_tests scripts
