+++ modules/phase_field/run_tests
@@ -9,20 +9,14 @@ app_name = 'phase_field'
 
 MODULE_DIR = os.path.abspath('..')
 MOOSE_DIR = os.path.abspath(os.path.join(MODULE_DIR, '..'))
-FRAMEWORK_DIR = os.path.abspath(os.path.join(MOOSE_DIR, 'framework'))
 #### See if MOOSE_DIR is already in the environment instead
 if os.environ.has_key("MOOSE_DIR"):
   MOOSE_DIR = os.environ['MOOSE_DIR']
-  FRAMEWORK_DIR = os.path.join(MOOSE_DIR, 'framework')
-if os.environ.has_key("FRAMEWORK_DIR"):
-  FRAMEWORK_DIR = os.environ['FRAMEWORK_DIR']
 
-# Import the TestHarness and Helper functions from the MOOSE toolkit
-sys.path.append(os.path.join(FRAMEWORK_DIR, 'scripts', 'common'))
-sys.path.append(os.path.join(FRAMEWORK_DIR, 'scripts', 'TestHarness'))
+sys.path.append(os.path.join(MOOSE_DIR, 'python'))
+import path_tool
+path_tool.activate_module('TestHarness')
 
-import tools
-
-# create an instance of the test harness passing in the main arg vector
-# and the name of the application
-tools.runTests(sys.argv, app_name, FRAMEWORK_DIR)
+from TestHarness import TestHarness
+# Run the tests!
+TestHarness.buildAndRun(sys.argv, app_name, MOOSE_DIR)
