+++ python/TestHarness/testers/CSVDiff.py
@@ -3,6 +3,7 @@ from CSVDiffer import CSVDiffer
 
 class CSVDiff(RunApp):
 
+  @staticmethod
   def validParams():
     params = RunApp.validParams()
     params.addRequiredParam('csvdiff',   [], "A list of files to run CSVDiff on.")
@@ -12,7 +13,6 @@ class CSVDiff(RunApp):
     params.addParam('delete_output_before_running',  True, "Delete pre-existing output files before running test. Only set to False if you know what you're doing!")
 
     return params
-  validParams = staticmethod(validParams)
 
   def __init__(self, name, params):
     RunApp.__init__(self, name, params)
