+++ python/TestHarness/testers/VTKDiff.py
@@ -5,6 +5,7 @@ from XMLDiffer import XMLDiffer
 
 class VTKDiff(RunApp):
 
+  @staticmethod
   def validParams():
     params = RunApp.validParams()
     params.addRequiredParam('vtkdiff',   [], "A list of files to exodiff.")
@@ -14,7 +15,6 @@ class VTKDiff(RunApp):
     params.addParam('delete_output_before_running',  True, "Delete pre-existing output files before running test. Only set to False if you know what you're doing!")
 
     return params
-  validParams = staticmethod(validParams)
 
   def __init__(self, name, params):
     RunApp.__init__(self, name, params)
